Refactors 2.0.0 code into a /v1/chatbot endpoint
